Move agreement tests to separate test file
This makes it easier for us to make the rest of the code independent
of ndarray.
